Massively compacted networking.
Waypoints are now always sent on join and automatically sends only accessible waypoints unless the player connecting is an admin.

WaypointsPayload is now the format, which sends a list of WaypointInfo which holds a Waypoint and a boolean indicating whether the waypoint is accessible. Unless the player is an admin, this value is always true